requirements.txt
absl-py==2.0.0
aiohttp==3.9.3
aiosignal==1.3.1
altgraph==0.17.4
asttokens==2.4.1
astunparse==1.6.3
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.3.1
click==8.1.3
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
cryptography==39.0.2
cx-Freeze==6.15.10
cx-Logging==3.1.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
discord==2.3.2
discord.py==2.3.2
distlib==0.3.8
et-xmlfile==1.1.0
executing==2.0.1
filelock==3.13.1
Flask==2.2.2
flatbuffers==23.5.26
fonttools==4.46.0
frozenlist==1.4.1
gast==0.5.4
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
idna==2.8
ImageHash==4.3.1
ipykernel==6.27.1
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.1
keras==2.15.0
keyboard==0.13.5
kiwisolver==1.4.5
libclang==16.0.6
lief==0.13.2
Markdown==3.5.1
MarkupSafe==2.1.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mcpy==2.0.0
ml-dtypes==0.2.0
mouse==0.7.1
MouseInfo==0.1.3
mss==9.0.1
multidict==6.0.5
nest-asyncio==1.5.8
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.2
pandas==2.0.0
parso==0.8.3
pefile==2023.2.7
Pillow==10.1.0
pip==24.0
platformdirs==4.1.0
plyer==2.1.0
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.7
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.17.2
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
PyMsgBox==1.0.9
pynput==1.7.6
pyOpenSSL==23.0.0
pyparsing==3.1.1
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytweening==1.0.4
pytz==2023.3
PyWavelets==1.4.1
pywin32==305
pywin32-ctypes==0.2.2
pyzmq==25.1.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.3
selenium==4.8.2
setuptools==65.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tk==0.1.0
tornado==6.4
tqdm==4.31.1
traitlets==5.14.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
virtualenv==20.25.1
wcwidth==0.2.12
websockets==8.1
Werkzeug==2.2.2
wheel==0.41.3
win10toast==0.9
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.4
